Changed the require to align with new {N} standard
This file earlier used the require
11 | var frame = require('ui/frame')
which has been deprecated, atleast in the latest version of {N} (6.3.2) as of writing (12 Apr, 2020) which causes an error message to be shown on running the app on the console:
JS: topmost() is deprecated. Use Frame.topmost() instead.
which I fixed by changing the require to
11 | var frame = require('tns-core'modules/ui/frame').Frame;
to align more with how the new module works.